Update testing-master.yml
U will see the following test do not pass because it cannot find the layout configuration.

- name: Run general script file for estimate bandwidth mode and ws dataflow
- name: Run general script file for user bandwidth mode and ws dataflow
- name: Run general script file for user bandwidth mode and is dataflow
- name: Run general script file for user bandwidth mode and os dataflow
The reason that without explicitly specifying the layout configuration through "-l", scalesim will use the default value set in the scale.py. which is "../layouts/conv_nets/test.csv", under the path where the existing scripts are launched, scalesim cannot find the layout file.

Note that, the reason why we set the default layout file as "../layouts/conv_nets/test.csv" is to pass the
"Run default SCALE-Sim configuration" configurations. So I change the workload configuration to run the scale.py at the path of root instead of under the 'scalesim' folder